Implemented the new parser function called hierarchySubtree.
This parser function is used to retrieve and display the subtree within a
hierarchy that is rooted at a specified row.

The version number is now 1.3.

Implemented the basics of a subhierarchy parser function.

Change-Id: I121b8cc7dc532da8e8292ccf9cf5638014a46b3b

Implemented the format parameter for the hierarchySubtree parser function.

Change-Id: I7db594e763ba69cb2dbcb4209c6e10620f2d1f15

Updated DOxygen documentation relating to the subhierarchy functionality.

Change-Id: I7f06909a1a547b64930893139ab61ea8075f5a1f

Updated DOxygen documentation regarding the hierarchySubtree parser function to include an additional sample invokation.

Change-Id: Ia3733ba3afaee718a226700df2736f5357fd1f04