Improved the stability of extension features.
Fixed a bug in hierarchySubtree to allow root elements to be displayed
when no root node is specified.

Fixed a bug in hierarchySubtree 'ul' display mode so that a new line is
automatically prepended to the output which allows the list to display
properly.

Modified hierarchyBreadcrumb to display information more clearly when the
previous node is the same as the parent node.

The version number is now 1.5.1.

Change-Id: Idd796da3f4d4b37e2c812c5789b6b3cb12a71d2c