Unwrap the results of the hierarchy tag function from the div that wraps it
A change in MW 1.30 causes Parser output to be wrapped in a div. This div causes the css selector in jstree to fail to find the li element for the hierarchy. By unwrapping the content from the div, the html structure becomes the same as before, allowing the css selector to succeed.

Change-Id: Id5bc89d781933b3e7402b111d04559896dca2226
(cherry picked from commit af0555808b72bfe8271026cf17676b48fa813a88)