Use DB_REPLICA instead of DB_SLAVE
Moved extension dependencies from extension.json to composer.json to
avoid test failure. This may introduce a dependency upon Semantic
MediaWiki 3.0.0, since I believe that is when full support for extension
registration was added to it. With earlier versions of Semantic
MediaWiki, the extension was not added to the extension list for
purposes of dependency checking. If support for earlier versions of SMW
is still desired, the line referencing Semantic MediaWiki will need to
be removed from composer.json.

Change-Id: I5e432bfe500d2bd02795e367cc4f083a0fecd8d2